// frontend/src/interfaces/user.ts
interface User {
    id: string;
    first_name: string;
    last_name: string;
    email: string;
    gender: string;
    age: string;
    country: string;
    occupation: string;
    profile_picture: string;
    username: string;
    password: string;
    phone_number: string;
  }
 
  export default User;